Removed duplicate results, but it's still pumping out non-sense...
added vector<LogicGate*>removeVectorDuplicates to avoid crazy large memory size required to run the program, however the current method to remove the duplicates is very inefficient and cpu demanding, there's a better way to do it but it's better to fix the problem where it's pumping out wrong results first.

I suspect that the problem might be that it's not just having 0 and 1, but also 2 to fill the void in the system.

However, I will try instead of filling the void with 2, just fill it with 0 to see if it will give correct result.

It's a pain to use F11 and break points to dissect my own project that I haven't touched since I got my job...